refactor: move getBaseColor into color util file
