+++ lib/components/form/batch-settings.tsx
@@ -24,7 +24,7 @@ import * as formActions from '../../actions/form'
 import { ComponentContext } from '../../util/contexts'
 import { generateModeSettingValues } from '../../util/api'
 import { getActiveSearch, hasValidLocation } from '../../util/state'
-import { getBaseColor, getDarkenedBaseColor } from '../util/base-color'
+import { getBaseColor, getDarkenedBaseColor } from '../util/colors'
 import { getFormattedMode } from '../../util/i18n'
 import { RoutingQueryCallResult } from '../../actions/api-constants'
 import { StyledIconWrapper } from '../util/styledIcon'
