+++ lib/components/util/colors.ts
@@ -1,6 +1,20 @@
 import colors from '@opentripplanner/building-blocks'
+import tinycolor, { Instance } from 'tinycolor2'
 
 const { blue, grey, red } = colors
+
+/**
+ * Finds the custom base color set in the configuration
+ */
+export const getBaseColor = (): string => {
+  return getComputedStyle(document.documentElement).getPropertyValue(
+    '--main-base-color'
+  )
+}
+
+export const getDarkenedBaseColor = (): Instance => {
+  return tinycolor(getBaseColor()).darken(10)
+}
 
 /**
  * A red color similar to Bootstrap's "alert" background #d9534f,
@@ -22,11 +36,14 @@ const BLUE_ON_WHITE = blue[800]
 
 const DARK_TEXT_GREY = '#333'
 
+const DEFAULT_ROUTE_COLOR = grey[800]
+
 export {
   grey,
   red,
   blue,
   DARK_TEXT_GREY,
+  DEFAULT_ROUTE_COLOR,
   RED_ON_WHITE,
   GRAY_ON_WHITE,
   BLUE_ON_WHITE
